Refactor Unpacker to support partial data
Previously, the unpacker did not handle partial data. But when writing
the `read_packed_from_stream` API I realized streams with silly buffers
would essentially be unsupported. I don't want to force a specific
sized buffer on users unlike with Packer (since we can just make our own
buffer for that case) and there isn't a way to really grow the buffer
without accidentally reading extra data that would be discarded.

So this reimplemented the Unpacker to support partial data. It still
tries to avoid bounds checking when possible, but will now always
attempt to read as much data as possible instead of avoiding partial
pieces of the stream.