Fix packing and add associated tests
I'll likely add more tests over time as I fill out certain features in
other places. As part of these tests, I had to fix a few bugs and add
some additional supporting APIs.

The Unpacker now doesn't immediately return if it receives an empty
output buffer. This is to allow passing in an empty buffer with some
remaining data to assert that no more data is in the input and no more
data needs to be written to the output.

This change is used in practice with `PackedStream::finish`. This new
function lets you do this empty output buffer check and get an error if
more data needs to be written.