Inline sharedshot into mpsc
Previously I believed sharedshot would serve as a base primitive for requests and channels. However, as time goes on, the needs of both have slightly changed so the sharedshot primitive gets inlined into the request and channels anyway so they can be customized.

So this just moves everything that was part of sharedshot into utility modules that can be reused, and sharedshot itself gets inlined into the mpsc channel since that was the last thing that used it directly. We'll customize this a bit more in the next few commit.