Add Lender type for easily iterating over List builders
A simple utility to help with iterating over List builders. Normally you'd
use a Range and write a manual for loop to access element builders, but it
doesn't work as well with other Iterators. Lender and ZippedLender provide
simple utilities to work with both. While they don't support Iterator
itself, we can still use `while let` with them which is close enough...