Add recapn-port library
This is the base library for a new method of generating "Plain Old Rust Types", hence the name "PORT".
Port is meant to be a single-pass de/serialization system more akin to serde vs the standard zero-copy
method Cap'n Proto libraries use. This makes the types easily movable and modifiable, but at the cost
that you lose backwards compability if you need to *reserialize*.

This is built on top of recapn, so you can use either one interchangably. They will have different code
generators though, which might make juggling a little complicated.

In this commit I also included some test files I've been using as a sketchpad for what the generated
code might look like. It's expected that it will be replaced with the final code generator output
before this is merged.