Fix reading empty objects on the end of a segment
SegmentReader::contains only works perfectly for the case where you want to read a non-empty object.
If an object was allocated to the end of the segment, then the start of the object will equal the end which
fails the check. But if an object is empty, this is valid.

This inlines the checks into the try_get_section function which already has comprehensive range checks.